backend/requirements.txt
fastapi
uvicorn
pandas
numpy
plotly
scipy
python-multipart
requests
openpyxl
motor
python-dotenv
matplotlib
seaborn
reportlab
httpx
tenacity